roles and granted authority

+++ src/main/java/br/com/forum_hub/domain/resposta/RespostaService.java
@@ -1,22 +1,26 @@
 package br.com.forum_hub.domain.resposta;
 
+import br.com.forum_hub.domain.perfil.HierarquiaService;
 import br.com.forum_hub.domain.topico.Status;
 import br.com.forum_hub.domain.topico.TopicoService;
+import br.com.forum_hub.domain.usuario.Usuario;
 import br.com.forum_hub.infra.exception.RegraDeNegocioException;
 import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.hierarchicalroles.RoleHierarchy;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class RespostaService {
+
     private final RespostaRepository repository;
     private final TopicoService topicoService;
-
-    public RespostaService(RespostaRepository repository, TopicoService topicoService) {
-        this.repository = repository;
-        this.topicoService = topicoService;
-    }
+    private final HierarquiaService hierarquiaService;
 
     @Transactional
     public Resposta cadastrar(DadosCadastroResposta dados, Long idTopico) {
@@ -37,8 +41,12 @@ public class RespostaService {
     }
 
     @Transactional
-    public Resposta atualizar(DadosAtualizacaoResposta dados) {
+    public Resposta atualizar(DadosAtualizacaoResposta dados, Usuario logado) {
         var resposta = buscarPeloId(dados.id());
+
+        if(hierarquiaService.usuarioNaoTemPermissoes(logado, resposta.getTopico().getAutor(), "ROLE_MODERADOR"))
+            throw new AccessDeniedException("Você não pode editar essa resposta!");
+
         return resposta.atualizarInformacoes(dados);
     }
 
@@ -47,10 +55,14 @@ public class RespostaService {
     }
 
     @Transactional
-    public Resposta marcarComoSolucao(Long id) {
+    public Resposta marcarComoSolucao(Long id, Usuario logado) {
         var resposta = buscarPeloId(id);
 
         var topico = resposta.getTopico();
+
+        if(!hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_INSTRUTOR"))
+            throw new RegraDeNegocioException("Você não pode marcar essa resposta como solução!");
+
         if(topico.getStatus() == Status.RESOLVIDO)
             throw new RegraDeNegocioException("O tópico já foi solucionado! Você não pode marcar mais de uma resposta como solução.");
 
@@ -58,10 +70,25 @@ public class RespostaService {
         return resposta.marcarComoSolucao();
     }
 
+    /*private boolean usuarioTemPermissoes(Usuario logado, Usuario autor) {
+        for(GrantedAuthority autoridade: logado.getAuthorities()){
+            var autoridadesAlcancaveis =  roleHierarchy.getReachableGrantedAuthorities(List.of(autoridade));
+
+            for(GrantedAuthority perfil: autoridadesAlcancaveis){
+                if(perfil.getAuthority().equals("ROLE_INSTRUTOR") || logado.getId().equals(autor.getId()))
+                    return true;
+            }
+        }
+        return false;
+    }*/
+
     @Transactional
-    public void excluir(Long id) {
+    public void excluir(Long id, Usuario logado) {
         var resposta = buscarPeloId(id);
         var topico = resposta.getTopico();
+
+        if(hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_MODERADOR"))
+            throw new AccessDeniedException("Você não pode apagar essa resposta!");
 
         repository.deleteById(id);
